implemented function to add note to trash or normalnotes collection

+++ src/app/note-list/note/note.component.ts
@@ -35,22 +35,28 @@ export class NoteComponent {
   }
 
   moveToTrash() {
-    if (this.note.id) {
+    if (this.note.id && this.note.type == 'note') {
       this.note.type = 'trash';
       let docId = this.note.id;
       delete this.note.id;
-      this.noteService.addNote(this.note, 'trash');
-      this.noteService.deleteNote('notes', docId);
+      this.noteService.addNote(this.note, "trash");
+      this.noteService.deleteNote("notes", docId);
     }
   }
 
   moveToNotes() {
-    this.note.type = 'note';
+    if (this.note.id && this.note.type == 'trash') {
+      this.note.type = 'note';
+      let docId = this.note.id;
+      delete this.note.id;
+      this.noteService.addNote(this.note, "notes");
+      this.noteService.deleteNote("trash", docId);
+    }
   }
 
   deleteNote() {
     if (this.note.id) {
-      this.noteService.deleteNote('trash', this.note.id);
+      this.noteService.deleteNote("trash", this.note.id);
     }
   }
